+++ src/fhops/productivity/__init__.py
@@ -59,6 +59,7 @@ from .harvester_ctl import (
     estimate_harvester_productivity_adv5n30,
     estimate_harvester_productivity_adv6n10,
     estimate_harvester_productivity_tn292,
+    estimate_harvester_productivity_kellogg1994,
 )
 from .kellogg_bettinger1994 import LoadType as KelloggLoadType
 from .kellogg_bettinger1994 import estimate_forwarder_productivity_kellogg_bettinger
@@ -153,6 +154,7 @@ __all__ = [
     "estimate_harvester_productivity_adv6n10",
     "estimate_harvester_productivity_adv5n30",
     "estimate_harvester_productivity_tn292",
+    "estimate_harvester_productivity_kellogg1994",
     "estimate_forwarder_productivity_small_forwarder_thinning",
     "estimate_forwarder_productivity_large_forwarder_thinning",
     "estimate_forwarder_productivity_kellogg_bettinger",